+++ src/lib/agent/plugins/vesu/schema/index.ts
@@ -1,0 +1,14 @@
+import { z } from 'zod';
+
+export const depositEarnSchema = z.object({
+  depositTokenSymbol: z
+    .string()
+    .describe("Symbol of the token to deposit (e.g., 'ETH', 'USDC')"),
+  depositAmount: z.string().describe('Amount of tokens to deposit'),
+});
+
+export const withdrawEarnSchema = z.object({
+  withdrawTokenSymbol: z
+    .string()
+    .describe("Symbol of the token to withdraw (e.g., 'ETH', 'USDC')"),
+});
